refactor(internal/transport/httpv1/users): BaseHandler has been move to the httpv1 module
